fix: msw in test setup file
