+++ packages/frontend/src/__tests__/loginPage.js
@@ -3,39 +3,6 @@ import userEvent from '@testing-library/user-event'
 import UserContext from '../context/User'
 import LoginPage from '../pages/loginPage/loginPage'
 
-import { rest } from 'msw'
-import { setupServer } from 'msw/node'
-
-const server = setupServer(
-    rest.get('https://tetsturl.invalidtld/*', (req, res, ctx) => {
-        console.log(req)
-        return res(
-            ctx.json({
-                test: 'test',
-            })
-        )
-    }),
-    rest.post('https://geth.testurl.invalidtld:8545/', (req, res, ctx) => {
-        // a geth request
-        return res(
-            ctx.json({
-                test: 'test',
-            })
-        )
-    })
-)
-
-beforeAll(() => {
-    // Establish requests interception layer before all tests.
-    server.listen()
-})
-
-afterAll(() => {
-    // Clean up after all tests are done, preventing this
-    // interception layer from affecting irrelevant tests.
-    server.close()
-})
-
 // mock needed for history and location hooks
 const mockHistoryPush = jest.fn()
